BackEnd/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["tcy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.4"
djangorestframework = "^3.14.0"
python-decouple = "^3.6"
django-cors-headers = "^3.13.0"


[[tool.poetry.source]]
name = "pypi-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false
secondary = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
